Fix SSD1306 build - remaining errors are from radiolib
